Switch to a custom List partition implementation.
The previous implementation from commons-collections could not split the
list as desired. The two possible options with this library are:
* The last partition is unbalanced to a potential partitionCount-1
elements more than all the other parititons.
* The number of partitions used may be one less than the requested
number of threads if an attempt is made to avoid the previous case (this
is the case prior to this commit).

This custom implementation ensures that all threads are used (unless the
number of features is less than the number of threads) and that the
difference in size between the smallest and largest partition is at most
one feature.
This also removes the dependency on the commons-collections4 library.